Revamp Main Menu state machine
