+++ Assets/Animations/UI/MainMenuController.controller
@@ -33,7 +33,7 @@ AnimatorController:
     m_IKPass: 0
     m_SyncedLayerAffectsTiming: 0
     m_Controller: {fileID: 9100000}
---- !u!114 &114290215792153718
+--- !u!114 &114499362236855942
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -41,15 +41,12 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: b80c61960de546e41a3f0aa142623d88, type: 3}
+  m_Script: {fileID: 11500000, guid: a918c683161303d4fb943d3548106806, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  defaultButtonNames:
-  - Play (Normal) Button
-  - Play (Hard) Button
-  - High Scores Button
-  - Quit Button
---- !u!114 &114478495272715844
+  objectNames:
+  - Default View
+--- !u!114 &114556229537306144
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -57,11 +54,37 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 4686e6f8c36d70d45849485ab0526dbf, type: 3}
+  m_Script: {fileID: 11500000, guid: a918c683161303d4fb943d3548106806, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  backButtonObjectName: Back Button
-  clearButtonObjectName: Clear Button
+  objectNames:
+  - High Scores View
+--- !u!114 &114626965248984166
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e14daf8bbdf4c204fb4bf48872f68abc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  objectNames:
+  - Default View
+--- !u!114 &114795626116659602
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e14daf8bbdf4c204fb4bf48872f68abc, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  objectNames:
+  - High Scores View
 --- !u!1101 &1101131140010825632
 AnimatorStateTransition:
   m_ObjectHideFlags: 1
@@ -122,7 +145,8 @@ AnimatorState:
   m_Transitions:
   - {fileID: 1101927603768108702}
   m_StateMachineBehaviours:
-  - {fileID: 114290215792153718}
+  - {fileID: 114499362236855942}
+  - {fileID: 114626965248984166}
   m_Position: {x: 50, y: 50, z: 0}
   m_IKOnFeet: 0
   m_WriteDefaultValues: 1
@@ -147,7 +171,8 @@ AnimatorState:
   m_Transitions:
   - {fileID: 1101131140010825632}
   m_StateMachineBehaviours:
-  - {fileID: 114478495272715844}
+  - {fileID: 114795626116659602}
+  - {fileID: 114556229537306144}
   m_Position: {x: 50, y: 50, z: 0}
   m_IKOnFeet: 0
   m_WriteDefaultValues: 1
